Commiting all files I used to annotate the course, using git.
Currently, I'm in the "Undoing Things" from the "Second Week" in the course "Introduction to Git and GitHub" on Coursera by Google.

I wish to make these notes as the files I will use as working examples of the course, instead of Python. These notes, and the historical process upon them, will serve as even fuller annotations of the progress of the course.